+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 12.8.1
 - Fix Advanced Search "No Value" for custom date and number fields in CustomFields.fnc.php
 - Fix SQL error escape CSP string in Content_Security_Policy/includes/common.fnc.php
 - Multilingual course title: use ParseMLField() in StudentList.php
+- Fix valid domain check: remove FILTER_VALIDATE_DOMAIN, too permissive in Content_Security_Policy/includes/common.fnc.php
 
 Changes in 12.8
 ---------------